i want you to send the message in extended manner
for example : if the prompt is given as " ask the manager to submit the project files", it should search for the manager and to send him the email - "Respected sir/madam, Submit the project files. Thank you!      Best regards..." - change the code accordingly likewise extract the role and the message to be informed for the voice prompt  and send the email accordingly